+++ sandbox/pyproject.toml
@@ -54,6 +54,9 @@ snowflake-connector-python = "^3.16.0"
 pyarrow = "^18.1.0"
 cryptography = "^45.0.5"
 databricks-sql-connector = "^4.0.5"
+nbclient = "^0.10.2"
+jupyter-client = "^8.6.3"
+ipykernel = "^6.29.5"
 
 [tool.poetry.group.dev.dependencies]
 pytest = ">=7.0.0"
